rename parse-ctl.R
parse-ctl.R has two exported functions: parse_ctl() with the main
logic and read_ctl(), a wrapper around the former.  Rename the source
file after the wrapper because 1) read_ctl() is more likely to be
called by an user and 2) an upcoming commit will add a write-ctl.R, so
it's nice to use read-ctl.R for symmetry.