record_parser: account for repeated option names
The IGNORE option in $DATA can be repeated.  I'm not sure how many
other options can be, but I don't think we can determine that clearly
from the documentation and, even if we could, checking for duplicate
option names is not the sort of error checking I think nmrec should
handle, at least for the record_parser level.

Deal with this by marking options in the template with their integer
index instead.  If we end up wanting to modify the template order,
integers may be more cumbersome to deal with, but 1) it's not clear to
me that we'll do that and 2) I'd rather layer helpers on top of the
integer representation than do something like encoding the count in
the name and/or making the options (optionally?) nested.