record: use method for parser function
To support combining multiple records for a given type, a record will
need to keep track of the previous record, if any, of its type.
That's one more field to pass into the parse_*_record() functions, so
just convert these into a method that has access to the record
environment.