add initial support for parsing combined record
All record types besides estimation, table, and scatterplot support
combining records.  nmrec currently has a parser for one record type
that supports combining across records, data.  For example, NM-TRAN
treats

  $data foo.csv
  $data ignore=C

the same as

  $data foo.csv ignore=C

To add support for this, teach parse_data_record() to check if there
is a previous data record and parse it if needed before moving to the
current record.