define first-level parser for theta records
This is the most complicated parser so far, but a good amount of this
code will be reused for the omega and sigma parser.  The
implementation feels a bit clunky.  Hopefully it can be cleaned up a
bit, but the tests should at least provide a good reference point.

The goal here is to parse the record to an object that entirely
captures the format on disk.  For example, "(0, 7.0)x5" is parsed to
an object that has an "x" option rather than expanding the value at
this point.  Also, the values are left as strings.

For cases where the elements are needed in a more processed form
(e.g., a numeric vector) the plan is to add another lay of helpers
that use these first-level objects as input.