tests: document multi-line spanning value within parens
As described in the next commit, nmrec doesn't support putting the
option value on the next line in general.  Add a test case for the one
exception.

As shown by this test case, the handling of trailing comments isn't
ideal, but it may not be worth the effort to address.