tstring: make size arguments optional
The current tstring users have pretty good guesses for the max size,
but that won't be true for the upcoming param parsers.

Set defaults for the sizes.  These are hopefully sensible for most
param values, but, as of the last commit, re-allocation is handled
better if these sizes are exceeded, so it matters less whether there
is a good guess upfront.