define first-level parser for omega and sigma records
This follows largely the same approach as the theta parsing from the
previous commit, and uses some shared helpers (defined in
parse-param.R).  The specification for omega and sigma records is the
same, so they share the same implementation.

As discussed in the previous commit, this is "first-level" in the
sense that the parsed objects capture and reflect the representation
on disk.  Higher-level helpers will be responsible for processing the
values further (e.g., creating a matrix object).