rework template class to use a single list
record_parser and the parsing for parameter values use a template
object that consists of two lists, `template` and `values`.  `values`
contains the nmrec_option objects, and `template` includes all other
objects, with integers pointing to a positions in `values`.

This works, but keeping two lists is unnecessary and complicates both
the code and test cases.  Instead just store all of the values as is
in a single list.  In cases where a list restricted to the
nmrec_option objects is needed, it's easy enough to filter the list
down to those.