record_parser: move process_options elsewhere
The record_parser object is used by every parse function and is one of
the most complicated parts of nmrec.  Most of its logic is concerned
with the lower-level functionality for managing the source list
(elems) or the destination list (lstr); it isn't coupled to the
specific record type.

Aside from name_raw, which is just used for more informative error
messages, the exception is the process_options() method.  That
involves knowing the option types, names, and abbreviations for a
given record.  Drop the method, and split the top-level function off
into its own file.