standardize return value for record_parser-modifying helpers
Several helper functions work by modifying the record_parser object
and don't return a value that's of interest.  In this case, return the
record_parser object invisibly, following the convention used for R6
methods.

(The main motivation for R6 objects doing that is chaining; there
aren't any spots in nmrec where this matters, so this change is just
for consistency.)