find_closing_paren: check if current item is an opening paren
find_closing_quote() returns 0L if it's not on an opening quote, while
find_closing_paren() just searches for the closing paren.  That
difference stems mostly from find_closing_quote() needing to look at
the opening quote to know whether to search for a single or double
quote.

So, in isolation, the find_closing_paren() behavior makes sense (it
doesn't really need to do the check), but it's confusing to have it
behave differently than find_closing_paren() given the similar purpose
and names.

Align the find_closing_paren() behavior with find_closing_quote(), and
adjust some callers so that they don't yank the opening paren before
find_closing_paren() is called.  (Also, in spots where it seems
readable enough, switch callers to using the find_closing_paren()
return value to detect if the current element is an opening paren.)