record_parser: fix handling of non-continuation '&'
gobble() checks if the current element is an ampersand and treats it
as a continuation if the next element is a linebreak (optionally with
preceding whitespace).  That works to gobble a continuation, but it
doesn't properly handle the "not a continuation" case.  For example,
IV.pdf gives this as an example of '&' as part of a value:

  FORMAT=s1PE15.8:160& ;

The current processing will correctly not consider this a
continuation, but it will fail to append '&' to the format value.

To handle ampersands that are part of a value, teach $yank() to check
whether the next element is an ampersand.