param: check that parameter values look numeric
The param parsers are pretty lax about what they'll take for parameter
values as long as the item is in the expected position and not a known
option.  Although most spots in nmrec don't try to validate option
values, do a simple check here that the value looks like a number to
avoid the param parser consuming lots of invalid (or at least
unsupported) syntax.