document main objects
This covers the main objects exposed to users: the nmrec_ctl_records
S3 class and the record and option R6 objects.  There are also
nmrec_element objects that should be documented at some point, but
those are more straightforward and users are less likely to want to
manipulate them.

Note that the record and option objects aren't exported at the moment,
though we'll want to do that eventually (or expose helpers for
constructing them) to support creating fresh records in addition to
modifying the returned records.