record, option: make name field read-only
The name field of the record and option objects is the resolved name.
This needs to be exposed to users so that they can reliably grab
objects for a given name regardless of the raw name used.  However,
this field should never be modified.

Make these read-only by converting them to active fields.