make check: error on any note
The check is note-free at the moment, so signal an error if anything
pops up.  (This will be used on Drone too.)

The "future file timestamps" check has an optional part where it hits
an external api to verify the system clock.  At the moment, that
service is down, and it has been unreliable in the past; trust the
system clock to avoid occasional failures due to the "unable to verify
current time" note.