record_parser: drop inaccessible code
should_absorb_amp() checks whether the previous position is a quote or
closing paren to avoid absorbing ampersands after cases like

  file='foo bar'&

and

 IGNORE=(AMT.EQN.1.0)&

(Based on NM-TRAN, these are considered continuations.)

All quote and paren handling yanks to the closing element;
should_absorb_amp() isn't triggered in those cases so there's no need
for this internal helper to consider that case.

The tests already cover the quote case.  Add a case for a continuation
after a closing paren.