get_record_option: return NULL if option isn't found
An error is signaled when an option isn't found in a record.  When
callers require the option be set, this behavior allows them to avoid
doing the error handling.  On the other hand, callers checking whether
a non-essential option might be present need to catch the error if
they want to use get_record_option().

At this point, it's not clear which case will be more common, but
switch to the NULL values to make processing a bit more concise for
the "not required" case, as suggested by @seth127 and @barrettk.