get_record_option: downcase non-flag/value names
When get_record_option() doesn't find the user-specified name in the
record's *_option_names map (i.e a name for an option_flag or
option_name option), it falls back to using the supplied name as is
for the resolved name, allowing the selection of an option_pos or
option_record_name option.  So, unlike NONMEM-based flag and value
options, these options require the full string, lower case.

This distinction is unlikely to be obvious to callers.  Loosen this
requirement a bit and accept any casing for option_pos and
option_record_name names, bringing the treatment more inline with the
flag/value options.

Thanks to @seth127 for the suggestion.