define parser for problem records
$PROBLEM records consist only of problem text that, if present, must
be on the first line, so the parser is fairly straightforward.
However, there are still a few details worth noting:

 * a semicolon in the first line does not signal a comment; it's just
   part of the problem text.

 * the record may have trailing comments or non-text lines.

 * the record may have no text at all, just the record name.  And in
   this case, there may not be whitespace after the record name.

For the "no text" case, add an empty string positional option if there
is whitespace after the record name.  That allows the user to grab the
option (e.g., with get_record_option) and modify it.  However, don't
add the option if there isn't trailing whitespace, because a user
grabbing and modifying the option in that case would just lead to an
invalid record name.

Closes #8