set_param: allow overriding default format specifier
The values for the set_* functions will typically be floating point
numbers (often the final estimates from a previous NONMEM run).  The
default as.character() behavior uses 15 significant digits, which is
too many in the context of initial estimates.

For this reason, param_set_value() formats the value with "%.3G".
That's probably good for most cases, but expose the format string as
an option to let callers override it.