set_{omega,sigma}: support resetting representation
A common source of values for the set_* function will be the final
estimates from another model.  Regardless of how the model for that
run specified OMEGA/SIGMA, NONMEM outputs variances and covariances,
so care must be taken to not feed those values into a control stream
that specifies something other than variance/covariance.

Add an option that lets the caller tell set_omega() and set_sigma()
that options for non-default representations should be ejected from
the modified records/options.

Go with two modes rather than a boolean option because it's not clear
whether we'll end up needing to support more variants.