set_param: drop parens if option was reduced to single item
Both discarding theta bounds ("(low,init,up)" => "(up)") and dropping
alternative representations from matrix options (e.g., "(SD 1)" =>
"(1)") can lead to a result that no longer needs parentheses.

In general, the set_* functions try to retain all of the original
control stream formatting when updating the init values.  However, in
the cases above, we're already dropping values at the request of a
non-default option, so it seems reasonable to simplify the result.

Note that the parentheses as left as is when updating a single initial
estimate surrounded by parentheses.