README: drop paragraph about higher-level forms
The motivation for "getting things in and out of higher-level forms"
was the functionality now covered by the new set_{theta,omega,sigma}
functions.  Those didn't end up needing to construct higher-level
forms from a record, and instead just rely on a map between parameter
space and the record options.