parse_ctl: abort if input contains newline
parse_ctl() accepts _lines_ of the control stream as input.  If a
caller instead passes the content of the control stream as a single
string and there are no newline characters before the problem record,
everything downstream is parsed as part of the problem record.  (If
the first line isn't a problem record, parse_ctl() aborts saying that
no records were found.)

Abort if any input line contains a newline character to make it
clearer that the input is invalid.

Re: https://github.com/metrumresearchgroup/nmrec/issues/13#issuecomment-1749436883