select_records: don't fail on string records
As described in the previous commit, the user may insert a string into
the $records list.  If select_records() is called after that point, it
fails because it assumes each $records item is an nmrec_record object
with a name field.

For a plain string, there should be no expectation that
select_records() can pull it out [*], but select_records() shouldn't
fail.  Make select_records() skip the string case.

[*] ... though at some point it may be worth considering a convenience
    function to convert $records strings to record objects.