add functions to extract theta, omega, and sigma init values
bbr is adding a feature to jitter the initial estimates from a control
stream.  To do so, it could loop through the initial estimate options
and jitter each one, but it's probably more convenient to be able to
work with the values in their native form (i.e. a numeric vector for
THETA or square matrix for OMEGA/SIGMA).

Note that naming these new functions get_{theta,omega,sigma} would
align nicely with the existing set_{theta,omega,sigma} functions, but
the get_* names would collide with bbr functions if a user attaches
both bbr and nmrec.

Re: #29