extract_param: support marking presence of specified flags
For jittering, bbr wants to know whether a value is fixed.  We could
add a dedicated option to return that information, but it seems likely
that at some point bbr or some other caller will want to know about
some other flag (e.g., UNINT or SD).

Add more general support for attaching a value that indicates if the
specified flag is present, allowing any known flag to be specified.

We could generalize this even further by allowing non-flag options
(e.g., returning the option value rather than FALSE/TRUE).  However,
THETA's NUMBERPOINTS=n seems like the only option this would be
relevant for, so it's not clear it's worth the trouble.