extract_theta: support getting bounds
bbr's jitter feature needs to account for the THETA bounds.  Add an
option that says to return the lower or upper bound values instead of
the initial estimates.

Another option would be to let the caller use a single call to return
some combination of init, low, and up values, but avoid doing that to
keep the return value consistent.