matrix parser: support SCALE option added in NM 7.6
As of NONMEM 7.6, $OMEGA and $SIGMA records can contain a SCALE
option.  Parsing it is relatively involved because it can occur in
many locations and be interspersed between other options and labels.

Leave one form unsupported: a diagonal record where SCALE is within
other init value options.  For example, in

  $OMEGA 0.1 SD 0.3

SD is an option for 0.1.  NM-TRAN allows

  $OMEGA 0.1 SCALE(2) SD 0.3

where SCALE(2) applies to 0.3 while SD still applies to 0.1.

That ordering is 1) confusing because SCALE applies to the following
values but is before an option that applies to the preceding value and
2) it is difficult for nmrec to both parse and represent.

At least for now, don't allow that form.  It's easy and clearer for
the user to instead write

  $OMEGA 0.1 SD SCALE(2) 0.3

The unsupported form will error with

  param_append_num_opt(lstr, "init", rp$yank())`: Expected numeric
  value for init, got "SD"

Unfortunately that's not very informative, but I haven't come up with
a straightforward way to give a better error message.

Re: #43