{set,extract}_matrix: warn if SCALE record is present
The SCALE option (new in NONMEM 7.6) for $OMEGA and $SIGMA records
specifies a scale factor that applies for all subsequent values (or
until the next SCALE option).

This option is likely to lead to a mismatch between what callers of
extract_{omega,sigma} and set_{omega,sigma} mean to specify and how
NONMEM interprets the values.

Flag the potential issue by signaling a warning.

At some point we may want to consider incorporating the SCALE option
in the value returned by extract_{omega,sigma} (similar to what can be
done with flags) so that callers can easily apply the scale factor to
the result.