process options: rearrange code to clarify handling of parentheses
parse_option_value handles a value flanked by parentheses regardless
of whether there is a non-empty separator between the option name and
the start of the value.  The initial "on open paren?" condition is
just used to determine if the separator should be empty.  However, at
a glance, a reader is likely to be confused that find_closing_paren is
called outside of the "on open paren?" condition.

As of the last commit, parse_option_sep can handle the empty separator
case, so rearrange things to call it unconditionally upfront.  Then,
to decide whether a parenthesized value needs to be yanked, prefer the
paren_open condition over checking the find_closing_paren return value
because understanding the latter requires the reader to know details
about how find_closing_paren behaves.