process options: tighten logic for ingesting value
Positioned after the option name, parse_option_sep finds the first
position after the separator by searching for the first item that's
not a whitespace or equal sign.  It errors if there isn't at least one
item for the separator.

That fails to consider that the first item after the separator may be
the start of a comment or a linebreak, neither of which should be
taken as the value of the option.

Make parse_option_sep error if the following item looks like a comment
or linebreak.  Also reworking it to handle the empty separator case.

Although not the main motivation, this new approach adds support for
any "stuck" option value as long as the option name and value are
broken apart by split_to_elements.  For example, NM-TRAN takes
file'foo' as the file option with value foo, and nmrec now accepts
that.  On the other hand, using print5 to specify the print option
with a value of 5 is still unsupported.