ci: update R version matrix
The jobs for R 4.2.3 and 4.3.1 matched the two most recent R versions
on Metworx at the time they were added.  Update them to match the two
most recent R versions on the latest Metworx AMI (25-02.01.00).

Drop the 4.1.3 run because it's unlikely to flag much that's not
covered by the runs for R 4.0.5 or the later versions.